fix(lens): move per-host activity-window facade out of the vocab leaf
activityWindow.ts drops its activeScope import and the activityWindow/setActivityWindow/activityWindowThresholdMs facade into a new terminal/activityWindowFilter.ts (mirroring showSleeping.ts), so the least-volatile vocabulary leaf no longer cycles back into the per-host owner.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-3, raised by lowy). Not pushed or merged.
